Refactored field style
