+++ ui-core/src/main/kotlin/com/processout/sdk/ui/core/component/field/POField.kt
@@ -18,8 +18,8 @@ import androidx.compose.ui.unit.dp
 import com.processout.sdk.ui.core.annotation.ProcessOutInternalApi
 import com.processout.sdk.ui.core.component.POBorderStroke
 import com.processout.sdk.ui.core.component.POText
-import com.processout.sdk.ui.core.style.input.POInputStateStyle
-import com.processout.sdk.ui.core.style.input.POInputStyle
+import com.processout.sdk.ui.core.style.field.POFieldStateStyle
+import com.processout.sdk.ui.core.style.field.POFieldStyle
 import com.processout.sdk.ui.core.theme.ProcessOutTheme
 
 /** @suppress */
@@ -71,22 +71,23 @@ object POField {
         }
 
     @Composable
-    fun custom(style: POInputStyle) = Style(
+    fun custom(style: POFieldStyle) = Style(
         normal = style.normal.toStateStyle(),
         error = style.error.toStateStyle()
     )
 
     @Composable
-    private fun POInputStateStyle.toStateStyle() = with(field) {
-        StateStyle(
-            text = POText.custom(style = text),
-            placeholderTextColor = colorResource(id = placeholderTextColorResId),
-            backgroundColor = colorResource(id = backgroundColorResId),
-            controlsTintColor = colorResource(id = controlsTintColorResId),
-            shape = RoundedCornerShape(size = border.radiusDp.dp),
-            border = POBorderStroke(width = border.widthDp.dp, color = colorResource(id = border.colorResId))
+    private fun POFieldStateStyle.toStateStyle() = StateStyle(
+        text = POText.custom(style = text),
+        placeholderTextColor = colorResource(id = placeholderTextColorResId),
+        backgroundColor = colorResource(id = backgroundColorResId),
+        controlsTintColor = colorResource(id = controlsTintColorResId),
+        shape = RoundedCornerShape(size = border.radiusDp.dp),
+        border = POBorderStroke(
+            width = border.widthDp.dp,
+            color = colorResource(id = border.colorResId)
         )
-    }
+    )
 
     val contentPadding: PaddingValues
         @Composable get() = PaddingValues(
