+++ ui-core/src/main/kotlin/com/processout/sdk/ui/core/style/field/POFieldStyle.kt
@@ -1,0 +1,10 @@
+package com.processout.sdk.ui.core.style.field
+
+import android.os.Parcelable
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+data class POFieldStyle(
+    val normal: POFieldStateStyle,
+    val error: POFieldStateStyle
+) : Parcelable
