--- ui-core/src/main/kotlin/com/processout/sdk/ui/core/style/input/POInputStyle.kt
@@ -1,10 +1,0 @@
-package com.processout.sdk.ui.core.style.input
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-
-@Parcelize
-data class POInputStyle(
-    val normal: POInputStateStyle,
-    val error: POInputStateStyle
-) : Parcelable
