+++ ui/src/main/kotlin/com/processout/sdk/ui/card/update/CardUpdateScreen.kt
@@ -85,7 +85,7 @@ internal fun CardUpdateScreen(
                 onEvent = onEvent
             )
             state.errorMessage?.let {
-                with(style.errorDescription) {
+                with(style.errorMessage) {
                     POText(
                         text = it,
                         color = color,
@@ -191,7 +191,7 @@ internal object CardUpdateScreen {
     data class Style(
         val title: POText.Style,
         val field: POField.Style,
-        val errorDescription: POText.Style,
+        val errorMessage: POText.Style,
         val actionsContainer: POActionsContainer.Style,
         val backgroundColor: Color,
         val dividerColor: Color,
@@ -203,10 +203,10 @@ internal object CardUpdateScreen {
         title = custom?.title?.let {
             POText.custom(style = it)
         } ?: POText.title,
-        field = custom?.input?.let {
+        field = custom?.field?.let {
             POField.custom(style = it)
         } ?: POField.default,
-        errorDescription = custom?.errorDescription?.let {
+        errorMessage = custom?.errorMessage?.let {
             POText.custom(style = it)
         } ?: POText.errorLabel,
         actionsContainer = custom?.actionsContainer?.let {
